requirements.txt
category_encoders==2.6.3
firebase_admin==6.5.0
geocoder==1.38.1
joblib==1.4.2
mysql_connector_repackaged==0.3.1
pandas==2.2.3
Pillow==10.1.0
Pillow==10.4.0
Pyrebase4==4.8.0
python-dotenv==1.0.1
scikit_learn==1.3.2
streamlit==1.35.0
streamlit_extras==0.4.7
streamlit_option_menu==0.3.13
